# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/",
    "paths": {
        "/api/icg/istioConfig": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Istio Resouce Config"
                ],
                "summary": "lists istio configurations of intio-gateway, virtual service, and destination rules.",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": ""
                    }
                }
            }
        },
        "/api/icg/lbConfig": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Istio Resouce Config"
                ],
                "summary": "defines load balance policy that applies to traffic intended for a service after routing has occurred.",
                "parameters": [
                    {
                        "description": "new LB policy to apply to service",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.lb"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    }
                }
            }
        },
        "/api/icg/weightConfig": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Istio Resouce Config"
                ],
                "summary": "defines weight policies that apply to traffic intended for a service after routing has occurred.",
                "parameters": [
                    {
                        "description": "new weights config",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.weights"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    }
                }
            }
        },
        "/example/helloworld": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "ping example",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.lb": {
            "type": "object",
            "properties": {
                "lb": {
                    "description": "round-robin",
                    "type": "integer",
                    "example": 4
                },
                "ns": {
                    "type": "string",
                    "example": "istio-test"
                },
                "svcname": {
                    "type": "string",
                    "example": "catalog"
                },
                "version": {
                    "type": "string",
                    "example": "v1"
                }
            }
        },
        "main.weights": {
            "type": "object",
            "properties": {
                "ns": {
                    "type": "string",
                    "example": "istio-test"
                },
                "svcname": {
                    "type": "string",
                    "example": "catalog"
                },
                "versions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "v1"
                    ]
                },
                "weights": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        100
                    ]
                }
            }
        }
    }
}